Feat: Module refactoring and renaming; Fix: Adaptations due to renaming (imports, README.md)
