+++ .gitignore
@@ -2,8 +2,8 @@
 
 read-write-file.rwf
 mcsim-traj.xyz
-MCSimulator/__pycache__/*
-MCSimulator/*/__pycache__/*
+MCsimulator/__pycache__/*
+MCsimulator/*/__pycache__/*
 
 
 # Byte-compiled / optimized / DLL files
