+++ MCSimulator/constants/constants.py
@@ -1,0 +1,2 @@
+IDEAL_GAS_CONST = 8.31446261815324
+KCAL_TO_J = 4184
