+++ mcsim.in
@@ -1,5 +1,5 @@
 parameter_file = params_mcsim.in; 
 restart_file = ar4_tetraeder.xyz;
 
-n_steps = 10000;
+n_steps = 2000;
 temp = 298.15;
